Code clean up. Added script to run integration tests.
