+++ tests/classes/encoded/data/JsonTest.php
@@ -32,38 +32,7 @@ class JsonTest extends PHPJsonUtilitiesTest
 
     public function setUp(): void
     {
-        $values = [
-            new Id(),
-            new Text(new Id()),
-            new ClassString(Id::class),
-            $this->randomChars(),
-            $this->randomClassStringOrObjectInstance(),
-            $this->randomFloat(),
-            $this->randomObjectInstance(),
-            [1, true, false, null, 'string', [], new Text($this->randomChars()), 'baz' => ['secondary_id' => new Id()], 'foo' => 'bar', 'id' => new Id(),],
-            true,
-            false,
-            function (): void {},
-            1,
-            1.2,
-            0,
-            [],
-            null,
-            'foo',
-            function (): void {},
-            json_encode("Foo bar baz"),
-            json_encode($this->randomChars()),
-            json_encode(['foo', 'bar', 'baz']),
-            json_encode([1, 2, 3]),
-            json_encode([PHP_INT_MIN, PHP_INT_MAX]),
-            new Directory(),
-            new Json($this->randomClassStringOrObjectInstance()),
-            new Json(json_encode(['foo', 'bar', 'baz'])),
-            new Reflection(new ClassString(Id::class)),
-            new ReflectionClass($this),
-            new ObjectReflection(new Id()),
-        ];
-        $data = $values[array_rand($values)];
+        $data = $this->randomData();
         $this->setExpectedJsonString($data);
         $this->setJsonTestInstance(
             new Json($data)
