+++ tests/dev/test/classes/TestClassB.php
@@ -2,6 +2,12 @@
 
 namespace Darling\PHPJsonUtilities\tests\dev\test\classes;
 
+/**
+ * This class is used by tests defined by the PHPJsonUtilities library.
+ *
+ * It should not be used in any other context.
+ *
+ */
 final class TestClassB
 {
