+++ tests/interfaces/encoded/data/JsonTestTrait.php
@@ -4,6 +4,7 @@ namespace Darling\PHPJsonUtilities\tests\interfaces\encoded\data;
 
 use Darling\PHPJsonUtilities\interfaces\encoded\data\Json;
 use \Closure;
+use \Darling\PHPJsonUtilities\classes\encoded\data\Json as JsonInsance;
 use \Darling\PHPJsonUtilities\tests\PHPJsonUtilitiesTest;
 use \Darling\PHPJsonUtilities\tests\interfaces\encoded\data\JsonTestTrait;
 use \Darling\PHPReflectionUtilities\classes\utilities\ObjectReflection;
@@ -59,7 +60,7 @@ trait JsonTestTrait
      *     $data = $values[array_rand($values)];
      *     $this->setExpectedJsonString($data);
      *     $this->setJsonTestInstance(
-     *         new Json($data)
+     *         new JsonInsance($data)
      *     );
      * }
      *
@@ -298,6 +299,42 @@ trait JsonTestTrait
                 'string __data__'
             )
         );
+    }
+
+    protected function randomData(): mixed
+    {
+        $values = [
+            new Id(),
+            new Text(new Id()),
+            new ClassString(Id::class),
+            $this->randomChars(),
+            $this->randomClassStringOrObjectInstance(),
+            $this->randomFloat(),
+            $this->randomObjectInstance(),
+            [1, true, false, null, 'string', [], new Text($this->randomChars()), 'baz' => ['secondary_id' => new Id()], 'foo' => 'bar', 'id' => new Id(),],
+            true,
+            false,
+            function (): void {},
+            1,
+            1.2,
+            0,
+            [],
+            null,
+            'foo',
+            function (): void {},
+            json_encode("Foo bar baz"),
+            json_encode($this->randomChars()),
+            json_encode(['foo', 'bar', 'baz']),
+            json_encode([1, 2, 3]),
+            json_encode([PHP_INT_MIN, PHP_INT_MAX]),
+            new Directory(),
+            new JsonInsance($this->randomClassStringOrObjectInstance()),
+            new JsonInsance(json_encode(['foo', 'bar', 'baz'])),
+            new Reflection(new ClassString(Id::class)),
+            new ReflectionClass($this),
+            new ObjectReflection(new Id()),
+        ];
+        return $values[array_rand($values)];
     }
 }
